# crc32c_arm64 performance optimization