@siying 2019-08-15
Thank you for your contribution. In the summary I didn't see the wall time comparison between the old and new approach. Does it run faster?