@Zhiwei-Dai 2019-08-17
@siying Thanks for your valuable replay.
My verification platform contains arm64 Kunpeng920(48Core@2.5G) and Kunpeng916(32Core@2.4G)

Analysization about Kunpeng920:
Much branch predictor optimazation work has been done on Kunpeng920
Theoretical bandwidth upper limit of linear crc32c:
bw = 8Bytes * (2.5 * 10^9)/s/3 = 6.67GB/s
Theoretical bandwidth upper limit of parallel optimization:
bw = 8Bytes * (2.5 * 10^9)/s = 20GB/s

Verification test:
a: db_bench --benchmarks="crc32c" on Kunpeng920
before: crc32c : 0.228 micros/op 4375470 ops/sec; 17091.7 MB/s (4096 per op)
after:    crc32c : 0.218 micros/op 4595390 ops/sec; 17950.7 MB/s (4096 per op)
speedup: +105.026%

b: crc32c_test
crc32c_test.cc tuning:
1,data size: BUFFER_SIZE
2,round times: StandardResults testcase forloop times
bw = size * round / time
speedup = bw(after) / bw(before)
On Kunpeng916:
            size   rounds   time(s)     bw(MB/s)    speedup
before 4M    60000    107.474    2233.098
after    4M    60000    66.557      3605.95      +161.48%
On Kunpeng920:
            size   rounds     time(s)             bw(MB/s)      speedup
before 4K    60000000  13.03293339   17983.288
after    4K    60000000  12.8695649     18211.571     +101.269%
before 1M   600000      41.10564898   14596.534
after    1M   600000      37.44392856   16023.960     +109.779%
before 4M   600000      168.4954331   14243.709
after    4M   600000      150.8937754   15905.229     +111.665%
before 64M 6000          37.06896454   10359.070
after    64M 6000          33.64855823   11412.079     +110.165%