@siying 2019-08-19
Regarding the theoretic analysis, can you explain why before the PR, we need to do "/3" but it's not needed after? If my understanding is correct, we are already doing a 3-way CRC computation even before the PR.