@Zhiwei-Dai 2019-08-20
@siying That's right. Before the PR, 3-way CRC computation has been used for coding.
The same: the PR's coding is 3-way CRC computation too, as line18~21
```
#define CRC32C24BYTES(ITR) \
    crc1 = crc32c_u64(crc1, *(buf64 + BLK_LENGTH + (ITR)));\
    crc2 = crc32c_u64(crc2, *(buf64 + BLK_LENGTH*2 + (ITR)));\
    crc1 = crc32c_u64(crc1, *(buf64 + (ITR)));
```
The difference: removing 'for' loop is good for branch predictor efficiency.
before the PR as line54~58
```
    for (int i = 0; i < BLK_LENGTH; i++, buf64++) {
      crc0 = crc32c_u64(crc0, *buf64);
      crc1 = crc32c_u64(crc1, *(buf64 + BLK_LENGTH));
      crc2 = crc32c_u64(crc2, *(buf64 + (BLK_LENGTH * 2)));
    }
```
the PR as line74~79:
```
    CRC32C7X24BYTES(0);
    CRC32C7X24BYTES(1);
    CRC32C7X24BYTES(2);
    CRC32C7X24BYTES(3);
    CRC32C7X24BYTES(4);
    CRC32C7X24BYTES(5);
```
from the verification test result, we can find that:
**smaller size data, less 'for' judgements, and the bw is more close to upper limit;
bigger size data, more 'for' judgements, and the PR can gain more speedup**