@siying 2019-08-20
I reread the data, so the 4K has 1.6% improvement, 1MB has 9.8% improvement. Is my understanding correct?