@Zhiwei-Dai 2019-08-21
@siying It's right. If the arm64 processer has a poor branch predictor, the speedup ratio will be higher.